Filter already added folders from auto-find projects
